pyproject.toml
[project]
name = "cifarcnn"
version = "0.1.0"
dependencies = [
    "torch>=2.7.0",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
]
requires-python = ">= 3.8"

[project.scripts]
"cifarcnn" = "cifarcnn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cifarcnn"]

[tool.pyright]
venvPath = "."
venv = ".venv"
